+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/plants/Sungrass.java
@@ -124,7 +124,7 @@ public class Sungrass extends Plant {
 
 		@Override
 		public float iconFadePercent() {
-			return Math.max(0, (target.HT - level) / target.HT);
+			return Math.max(0, (target.HT - level) / (float)target.HT);
 		}
 		
 		@Override
